+++ src/main/java/de/vette/idea/neos/lang/xliff/XliffParserDefinition.java
@@ -1,0 +1,56 @@
+package de.vette.idea.neos.lang.xliff;
+
+import com.intellij.lang.*;
+import com.intellij.lang.xml.XMLLanguage;
+import com.intellij.lexer.Lexer;
+import com.intellij.lexer.XmlLexer;
+import com.intellij.openapi.project.Project;
+import com.intellij.psi.FileViewProvider;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiFile;
+import com.intellij.psi.impl.source.parsing.xml.XmlParser;
+import com.intellij.psi.tree.IFileElementType;
+import com.intellij.psi.tree.TokenSet;
+import org.jetbrains.annotations.NotNull;
+
+public class XliffParserDefinition implements ParserDefinition {
+    @Override
+    public @NotNull Lexer createLexer(Project project) {
+        return new XmlLexer();
+    }
+
+    @Override
+    public @NotNull PsiParser createParser(Project project) {
+        return LanguageParserDefinitions.INSTANCE.forLanguage(Language.findInstance(XMLLanguage.class)).createParser(project);
+    }
+
+    @Override
+    public @NotNull IFileElementType getFileNodeType() {
+        return XliffFile.XLIFF_FILE_ELEMENT_TYPE;
+    }
+
+    @Override
+    public @NotNull TokenSet getCommentTokens() {
+        return LanguageParserDefinitions.INSTANCE.forLanguage(Language.findInstance(XMLLanguage.class)).getCommentTokens();
+    }
+
+    @Override
+    public @NotNull TokenSet getStringLiteralElements() {
+        return TokenSet.EMPTY;
+    }
+
+    @Override
+    public @NotNull TokenSet getWhitespaceTokens() {
+        return LanguageParserDefinitions.INSTANCE.forLanguage(Language.findInstance(XMLLanguage.class)).getWhitespaceTokens();
+    }
+
+    @Override
+    public @NotNull PsiElement createElement(ASTNode node) {
+        throw new IllegalArgumentException("Unknown element: "+node);
+    }
+
+    @Override
+    public @NotNull PsiFile createFile(@NotNull FileViewProvider viewProvider) {
+        return new XliffFile(viewProvider);
+    }
+}